requirements.txt
Django==1.9.6
django-ckeditor==5.0.3
django-debug-toolbar==1.4
django-el-pagination==2.1.2
Pillow==3.2.0
psycopg2==2.6.1
sorl-thumbnail==12.3
sqlparse==0.1.19
Unidecode==0.4.19
